backend/pyproject.toml
[project]
name = "synthetic-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "httpx>=0.27.0",
    "kubernetes>=28.1.0",
    "minio>=7.2.18",
    "pydantic>=2.12.3",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "langgraph>=1.0.7",
]
